requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
distlib==0.3.1
Django==2.2.9
filelock==3.0.12
idna==2.10
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
numpy==1.19.3
opencv-python==4.4.0.46
packaging==20.9
pluggy==0.13.1
py==1.10.0
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.2
pytest-django==3.8.0
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
toml==0.10.2
twilio==6.53.0
urllib3==1.26.3
virtualenv==20.1.0
virtualenvwrapper-win==1.2.6
wrapt==1.12.1
